+++ test/integration/lib.py
@@ -1,6 +1,9 @@
-from . import forge
-from forge_sdk import protos, utils
+import json
 from uuid import uuid4
+
+from forge_sdk import utils
+
+from test.integration import forge