+++ test/integration/test_delegate.py
@@ -1,17 +1,10 @@
 from time import sleep
 
-from forge_sdk import protos
-from test.integration import forge
-
-
-def create_users():
-    w1 = forge.create_wallet(moniker="alice", passphrase='abcd1234')
-    w2 = forge.create_wallet(moniker="bobby", passphrase='abcd1234')
-    w3 = forge.create_wallet(moniker="claire", passphrase='abcd1234')
-    return w1.wallet, w2.wallet, w3.wallet
-
+from forge_sdk import protos, utils
+from test.integration import forge, create_asset, create_users
 
 boss, helper, partner = create_users()
+asset = create_asset(partner)
 
 value = forge.to_unit(50)
 sleep(5)
@@ -39,6 +32,7 @@ def test_delegate():
     assert forge.fetch_balance(partner.address) - partner_bal == value
     assert forge.fetch_balance(helper.address) == helper_bal
 
+
 def test_revoke_delegate():
     ops = protos.DelegateOp(
             type_url='fg:t:transfer'
@@ -61,5 +55,32 @@ def test_revoke_delegate():
     assert res.code == 58
 
 
+def test_delegate_exchange():
+    ops = protos.DelegateOp(
+            type_url='fg:t:exchange'
+    )
+    asset_state = forge.fetch_asset(asset)
+    boss_bal = forge.fetch_balance(boss.address)
+    partner_bal = forge.fetch_balance(partner.address)
+
+    assert asset_state.owner == partner.address
+    res = forge.delegate(to=helper.address, ops=[ops], wallet=boss,
+                         commit=True)
+    assert res.code == 0
+
+    sender = utils.exchange_info(assets=[asset])
+    receiver = utils.exchange_info(value=value)
+    tx = forge.prepare_exchange(sender=sender, receiver=receiver,
+                                wallet=partner)
+    res = forge.finalize_exchange(tx=tx, wallet=helper,
+                                  delegatee=boss.address, commit=True)
+    assert res.code == 0
+
+    new_state = forge.fetch_asset(asset)
+    assert new_state.owner == boss.address
+    assert forge.fetch_balance(partner.address) - partner_bal == value
+    assert forge.fetch_balance(boss.address) + value == boss_bal
+
+
 if __name__ == '__main__':
-    test_revoke_delegate()
+    test_delegate_exchange()